fix(Storybook): moves ComponentStyleTable to components
